requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
pillow
tqdm
pyyaml
scikit-image
lpips
matplotlib
